# whatsup-client/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:9090",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/chats": {
            "get": {
                "tags": [
                    "chat"
                ],
                "operationId": "getChatsByReceiver",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChatResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "chat"
                ],
                "operationId": "createChat",
                "parameters": [
                    {
                        "name": "senderId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "receiverId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StringResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/messages/chats/{chatId}": {
            "get": {
                "tags": [
                    "messages"
                ],
                "operationId": "getMessages",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "senderId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "receiverId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "TEXT",
                                "IMAGE",
                                "VIDEO",
                                "AUDIO",
                                "DOCUMENT"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortDir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageMessageResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StringResponse": {
                "type": "object",
                "properties": {
                    "response": {
                        "type": "string"
                    }
                }
            },
...
            "MessageResponse": {
                "type": "object",
                "properties": {
                    "messageId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "content": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "TEXT",
                            "IMAGE",
                            "VIDEO",
                            "AUDIO",
                            "DOCUMENT"
                        ]
                    },
                    "state": {
                        "type": "string",
                        "enum": [
                            "SENT",
                            "SEEN"
                        ]
                    },
                    "media": {
                        "type": "string",
                        "format": "byte"
                    },
                    "senderId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "receiverId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PageMessageResponse": {
                "type": "object",
                "properties": {
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MessageResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "unpaged": {
                        "type": "boolean"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    }
                }
            },
            "ChatResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "unreadCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "lastMessage": {
                        "type": "string"
                    },
                    "lastMessageTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "senderId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "receiverId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "receiverOnline": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8080/realms/whatsup/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:8080/realms/whatsup/protocol/openid-connect/auth",
                        "scopes": {}
                    }
                }
            }
        }
    }
}