# whatsup-client/src/services/models/MessageResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MessageResponse
 */
export interface MessageResponse {
    /**
     * 
     * @type {string}
     * @memberof MessageResponse
     */
    messageId?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageResponse
     */
    content?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageResponse
     */
    type?: MessageResponseTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof MessageResponse
     */
    state?: MessageResponseStateEnum;
    /**
     * 
     * @type {string}
     * @memberof MessageResponse
     */
    media?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageResponse
     */
    senderId?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageResponse
     */
    receiverId?: string;
    /**
     * 
     * @type {Date}
     * @memberof MessageResponse
     */
    createdAt?: Date;
}


/**
 * @export
 */
export const MessageResponseTypeEnum = {
    Text: 'TEXT',
    Image: 'IMAGE',
    Video: 'VIDEO',
    Audio: 'AUDIO',
    Document: 'DOCUMENT'
} as const;
export type MessageResponseTypeEnum = typeof MessageResponseTypeEnum[keyof typeof MessageResponseTypeEnum];

/**
 * @export
 */
export const MessageResponseStateEnum = {
    Sent: 'SENT',
    Seen: 'SEEN'
} as const;
export type MessageResponseStateEnum = typeof MessageResponseStateEnum[keyof typeof MessageResponseStateEnum];


/**
 * Check if a given object implements the MessageResponse interface.
 */
export function instanceOfMessageResponse(value: object): value is MessageResponse {
    return true;
}

export function MessageResponseFromJSON(json: any): MessageResponse {
    return MessageResponseFromJSONTyped(json, false);
}

export function MessageResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessageResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'messageId': json['messageId'] == null ? undefined : json['messageId'],
        'content': json['content'] == null ? undefined : json['content'],
        'type': json['type'] == null ? undefined : json['type'],
        'state': json['state'] == null ? undefined : json['state'],
        'media': json['media'] == null ? undefined : json['media'],
        'senderId': json['senderId'] == null ? undefined : json['senderId'],
        'receiverId': json['receiverId'] == null ? undefined : json['receiverId'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function MessageResponseToJSON(json: any): MessageResponse {
    return MessageResponseToJSONTyped(json, false);
}

export function MessageResponseToJSONTyped(value?: MessageResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'messageId': value['messageId'],
        'content': value['content'],
        'type': value['type'],
        'state': value['state'],
        'media': value['media'],
        'senderId': value['senderId'],
        'receiverId': value['receiverId'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}