# whatsup-client/src/services/models/PageMessageResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PageableObject } from './PageableObject';
import {
    PageableObjectFromJSON,
    PageableObjectFromJSONTyped,
    PageableObjectToJSON,
    PageableObjectToJSONTyped,
} from './PageableObject';
import type { MessageResponse } from './MessageResponse';
import {
    MessageResponseFromJSON,
    MessageResponseFromJSONTyped,
    MessageResponseToJSON,
    MessageResponseToJSONTyped,
} from './MessageResponse';
import type { SortObject } from './SortObject';
import {
    SortObjectFromJSON,
    SortObjectFromJSONTyped,
    SortObjectToJSON,
    SortObjectToJSONTyped,
} from './SortObject';

/**
 * 
 * @export
 * @interface PageMessageResponse
 */
export interface PageMessageResponse {
    /**
     * 
     * @type {number}
     * @memberof PageMessageResponse
     */
    totalElements?: number;
    /**
     * 
     * @type {number}
     * @memberof PageMessageResponse
     */
    totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageMessageResponse
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageMessageResponse
     */
    last?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PageMessageResponse
     */
    size?: number;
    /**
     * 
     * @type {Array<MessageResponse>}
     * @memberof PageMessageResponse
     */
    content?: Array<MessageResponse>;
    /**
     * 
     * @type {number}
     * @memberof PageMessageResponse
     */
    number?: number;
    /**
     * 
     * @type {number}
     * @memberof PageMessageResponse
     */
    numberOfElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageMessageResponse
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {SortObject}
     * @memberof PageMessageResponse
     */
    sort?: SortObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageMessageResponse
     */
    empty?: boolean;
}

/**
 * Check if a given object implements the PageMessageResponse interface.
 */
export function instanceOfPageMessageResponse(value: object): value is PageMessageResponse {
    return true;
}

export function PageMessageResponseFromJSON(json: any): PageMessageResponse {
    return PageMessageResponseFromJSONTyped(json, false);
}

export function PageMessageResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PageMessageResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'totalElements': json['totalElements'] == null ? undefined : json['totalElements'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
        'first': json['first'] == null ? undefined : json['first'],
        'last': json['last'] == null ? undefined : json['last'],
        'size': json['size'] == null ? undefined : json['size'],
        'content': json['content'] == null ? undefined : ((json['content'] as Array<any>).map(MessageResponseFromJSON)),
        'number': json['number'] == null ? undefined : json['number'],
        'numberOfElements': json['numberOfElements'] == null ? undefined : json['numberOfElements'],
        'pageable': json['pageable'] == null ? undefined : PageableObjectFromJSON(json['pageable']),
        'sort': json['sort'] == null ? undefined : SortObjectFromJSON(json['sort']),
        'empty': json['empty'] == null ? undefined : json['empty'],
    };
}

export function PageMessageResponseToJSON(json: any): PageMessageResponse {
    return PageMessageResponseToJSONTyped(json, false);
}

export function PageMessageResponseToJSONTyped(value?: PageMessageResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'totalElements': value['totalElements'],
        'totalPages': value['totalPages'],
        'first': value['first'],
        'last': value['last'],
        'size': value['size'],
        'content': value['content'] == null ? undefined : ((value['content'] as Array<any>).map(MessageResponseToJSON)),
        'number': value['number'],
        'numberOfElements': value['numberOfElements'],
        'pageable': PageableObjectToJSON(value['pageable']),
        'sort': SortObjectToJSON(value['sort']),
        'empty': value['empty'],
    };
}